% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Challenge_dat.R
\docType{data}
\name{Challenge_dat}
\alias{Challenge_dat}
\title{Genetic, metabolite, and outcome data for the 2019 MR Conference Data Challenge.}
\format{A data frame with 150 rows and 267 variables. A full description of the data is available
by using the command \code{vignette("challenge")}.}
\source{
\itemize{
\item \url{http://www.computationalmedicine.fi/data#NMR_GWAS}
\item \url{http://lipidgenetics.org/}
\item \url{http://amdgenetics.org/}
\item \url{http://diagram-consortium.org/downloads.html}
\item \url{http://www.megastroke.org/}
}
}
\usage{
Challenge_dat
}
\description{
A dataset containing summary data for the estimated associations between 118 metabolite risk factors and
150 genetic variants, with metabolite information quantified using nuclear magnetic resonance (NMR) spectroscopy 
metabolomics. The data also includes information on 7 outcomes; age-related macular degeneration, alzheimers's disease,
 type 2 diabetes, Ischemic stroke, large artery stroke, cardioembolic stroke, and small vessel stroke.
}
\details{
Challenge_dat.
}
\author{
Wes Spiller; Jack Bowden; Verena Zuber.
}
\keyword{datasets}
