% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRChallenge2019-package.r
\docType{package}
\name{MRChallenge2019-package}
\alias{MRChallenge2019-package}
\title{MRChallenge2019:}
\description{
The package provides summary data for the estimated associations between 118 
metabolite risk factors and 150 genetic variants, with metabolite information 
quantified using nuclear magnetic resonance (NMR) spectroscopy metabolomics. 
The data also includes information on 7 outcomes as described below. A 
second data frame is also included providing the abbreviation, full name, 
overall heritability of the trait, and classification of each NMR trait. 
Finally, a document is included giving a description of the data and broad 
aims of the 2019 MR Conference Data Challenge.
}
