% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRA_dat.R
\docType{data}
\name{NMRA_dat}
\alias{NMRA_dat}
\title{A data frame containing the abbreviation, full name, overall heritability of the trait, and classification of each NMR trait.}
\format{A data frame with 118 rows and 21 variables. A full description of the data is available
by using the command \code{vignette("MRDataChallenge19")}.}
\source{
\itemize{
\item \url{https://nightingalehealth.com/biomarkers}
}
}
\usage{
NMRA_dat
}
\description{
A data frame containing the abbreviation, full name, overall heritability of the trait, and classification of the NMR trait.
The classification into categories has been done according to the Nightingale biomarker classification. Risk factor sets named XXX_in are 
derived from the name of specifc cholesterol measurements, e.g. Total_Cholesterol_in or Free_Cholesterol_in.
}
\details{
NMRA_dat.
}
\author{
Wes Spiller; Jack Bowden; Verena Zuber.
}
\keyword{datasets}
